Tema3/tasks/tasks/requirements.txt
Django==3.2.12
djangorestframework==3.14.0
gspread==6.1.0
oauth2client==4.1.3
gunicorn==21.2.0
google-cloud-storage==2.16.0
google-cloud-pubsub==2.21.1
django-cors-headers==4.3.1
